desktop/requirements-mac.txt
pywebview==5.1
pyinstaller==6.16.0
